requirements.txt
# Core dependencies
# Python 3.9+ required (not a pip package)

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0

# AI/ML
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99
accelerate>=0.24.0

# UI
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Optional (for faster downloads)
requests>=2.31.0
urllib3>=2.0.0

# Development (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
pytest>=7.4.0
black>=23.9.0
